improved velocity dispersion sampling
