+++ tests/test_Deflectors/test_all_lens_galaxies.py
@@ -2,8 +2,8 @@ from astropy.cosmology import FlatLambdaCDM
 from slsim.Deflectors.all_lens_galaxies import (
     AllLensGalaxies,
     fill_table,
-    vel_disp_abundance_matching,
 )
+from slsim.Deflectors.velocity_dispersion import vel_disp_abundance_matching
 from slsim.Pipelines.skypy_pipeline import SkyPyPipeline
 from astropy.units import Quantity
 from astropy.table import Table
