+++ tests/test_Deflectors/test_velocity_dispersion.py
@@ -80,6 +80,3 @@ def test_schechter_vdf():
         noise=True,
     )
     assert len(z_list) == 117
-
-    # plt.hist(np.log10(vel_disp_list))
-    # plt.show()
